Use explicit call_types when calling read_tsv
This suppresses the col_spec output and is generally better practice to
always specify the column types rather than relying on guessing if you
know the expected types.

In readr 2.0.0 the column types will be shown as a message, which causes
the tests against `expect_silent()` to fail.

In addition what used to be a warning is now an error when you pass the
wrong input type, so that change was made as well.