Possible fix for #8
Previously the description column filtered out comments (everything
following a '#' sign). This caused all entries in the search
domtableout#' sign). This caused all entries in the search domtblout
description column to be empty and caused the error seen in #8.

The solution is not entirely satisfactory. There may be cases were we
actually do want to ignore comments. Also, the search domtblout
description comment contains information on the hit that should possibly
be extracted and stored in dedicated columns.